Support encoding of 3+ UTF-8 byte characters
For such characters, the java `Char` which contains only the first 2
bytes cannot be converted to a valid string which result in a string
containing a single byte `0x3F`.

By iterating over the UTF-8 bytes instead of 2-byte characters, we can
handle all different cases, treating the allowed characters as the
single byte characters they are.